Move cluster-api/node/util to cluster-api/pkg/controller/noderefutil
This moves the node ref util helpers to where they are actually used.

Update machineset/status.go to use helpers from noderefutil rather
duplicating the logic.
