+++ pkg/controller/noderefutil/util_test.go
@@ -1,4 +1,4 @@
-package util
+package noderefutil
 
 import (
 	"reflect"
